+++ packages/core/src/StateNode.ts
@@ -96,7 +96,8 @@ import {
   getChildren,
   getAllStateNodes,
   isInFinalState,
-  isLeafNode
+  isLeafNode,
+  getTagsFromConfiguration
 } from './stateUtils';
 import { createInvocableActor } from './Actor';
 import { toInvokeDefinition } from './invokeUtils';
@@ -252,6 +253,8 @@ class StateNode<
 
   public __xstatenode: true = true;
 
+  public description?: string;
+
   private __cache = {
     events: undefined as Array<TEvent['type']> | undefined,
     relativeValue: new Map() as Map<StateNode<TContext>, StateValue>,
@@ -323,6 +326,7 @@ class StateNode<
       ? this.machine.schema
       : (this.config as MachineConfig<TContext, TStateSchema, TEvent>).schema ??
         ({} as this['schema']);
+    this.description = this.config.description;
 
     if (!IS_PRODUCTION) {
       warn(
@@ -519,7 +523,9 @@ class StateNode<
       meta: this.meta,
       order: this.order || -1,
       data: this.doneData,
-      invoke: this.invoke
+      invoke: this.invoke,
+      description: this.description,
+      tags: this.tags
     };
   }
 
@@ -712,7 +718,8 @@ class StateNode<
       ...state,
       value: this.resolve(state.value),
       configuration,
-      done: isInFinalState(configuration, this)
+      done: isInFinalState(configuration, this),
+      tags: getTagsFromConfiguration(configuration)
     });
   }
 
@@ -1339,8 +1346,8 @@ class StateNode<
     // Preserve original history after raised events
     maybeNextState.history = history;
 
-    maybeNextState.tags = new Set(
-      flatten(maybeNextState.configuration.map((sn) => sn.tags))
+    maybeNextState.tags = getTagsFromConfiguration(
+      maybeNextState.configuration
     );
 
     return maybeNextState;